1. rollback init polaris-context to application phase 2. set example polaris server address to 127.0.0.1
