+++ CHANGELOG.md
@@ -14,5 +14,6 @@
 - [feat:optimize metadata module.](https://github.com/Tencent/spring-cloud-tencent/pull/70)
 - [feat: optimize pom dependency and demo](https://github.com/Tencent/spring-cloud-tencent/pull/71)
 - [feat:upgrade polaris version to 1.3.0.](https://github.com/Tencent/spring-cloud-tencent/pull/72)
+- [feat:rollback init polaris-context to application phase](https://github.com/Tencent/spring-cloud-tencent/pull/73)
