+++ spring-cloud-tencent-examples/polaris-circuitbreaker-example/polaris-circuitbreaker-example-b/src/main/resources/bootstrap.yml
@@ -5,6 +5,6 @@ spring:
     name: polaris-circuitbreaker-example-b
   cloud:
     polaris:
-      address: grpc://159.75.195.18:8091 # The demo address officially provided by Polaris can be accessed directly http://159.75.195.18
+      address: grpc://127.0.0.1:8091
       namespace: default
 is-throw-runtime-exception: false
