+++ spring-cloud-tencent-examples/polaris-discovery-example/discovery-callee-service/src/main/resources/bootstrap.yml
@@ -5,7 +5,7 @@ spring:
     name: DiscoveryCalleeService
   cloud:
     polaris:
-      address: grpc://159.75.195.18:8091 # The demo address officially provided by Polaris can be accessed directly http://159.75.195.18
+      address: grpc://127.0.0.1:8091
       namespace: default
 #    consul:
 #      port: 8500
