--- spring-cloud-tencent-examples/polaris-discovery-example/discovery-caller-service/src/main/resources/polaris.yml
@@ -1,6 +1,0 @@
-global:
-  system:
-    discoverCluster:
-      sameAsBuiltin: true
-    healthCheckCluster:
-      sameAsBuiltin: true
