+++ spring-cloud-tencent-examples/polaris-gateway-example/gateway-scg-service/src/main/resources/bootstrap.yml
@@ -12,7 +12,7 @@ spring:
         transitive:
           - a
     polaris:
-      address: grpc://159.75.195.18:8091 # The demo address officially provided by Polaris can be accessed directly http://159.75.195.18
+      address: grpc://127.0.0.1:8091
       namespace: default
     gateway:
       discovery:
